changed shamba refactors

+++ app/Traits/RoomTrait.php
@@ -12,19 +12,24 @@ use Livewire\WithPagination;
 trait RoomTrait {
     use WithPagination;
     public function getTotalRooms(){
-        return Room::get()->count();
+        return Room::where('team_id', auth()->user()->currentTeam->id)->get()->count();
     }
 
     public function getAllBookedRooms(){
-        return Room::where('is_available', 0)->with('staff')->with('room_types')->get();
+        return Room::where('is_available', 0)->where('team_id', auth()->user()->currentTeam->id)->orderByDesc('created_at')->with('staff')->with('room_types')->get();
     }
 
     public function getAllRooms($team_id){
-        return Room::with('staff')->where('team_id', $team_id)->with('room_types')->get();
+        return Room::with('staff')->where('team_id', $team_id)->orderByDesc('created_at')->with('room_types')->get();
     }
 
+    // For API use
     public function getAllRoomTypes2($team_id){
-        return RoomType::with('users')->where('team_id', $team_id)->get();
+        return RoomType::with('users')->where('team_id', $team_id)->orderByDesc('created_at')->get();
+    }
+
+    public function getAvailableRooms(){
+        return Room::where('is_available', 1)->where('team_id', auth()->user()->currentTeam->id)->orderByDesc('created_at')->with('staff')->with('room_types')->get();
     }
     
     // public function getAllRoomTypes(){
@@ -32,7 +37,7 @@ trait RoomTrait {
     // }
 
     public function getRoomType($id){
-        return RoomType::where('id', $id)->get()->first();
+        return RoomType::where('id', $id)->where('team_id', auth()->user()->currentTeam->id)->orderByDesc('created_at')->first();
     }
 
     public function saveRoomType($request){
@@ -83,7 +88,7 @@ trait RoomTrait {
     }
 
     public function toggleRoomStatus($id){
-        $room = Room::find($id);
+        $room = Room::where('id', $id)->where('team_id', auth()->user()->currentTeam->id)->first();
         $room->is_available = !$room->is_available;
         $room->save();
     }
@@ -91,8 +96,19 @@ trait RoomTrait {
     public function removeRoomType($id){
     
         try {
-            DB::table('room_types')->where('id', '=', $id)->delete();
-            DB::table('rooms')->where('room_types_id', '=', $id)->delete();
+            DB::table('room_types')->where('id', '=', $id)->where('team_id', auth()->user()->currentTeam->id)->delete();
+            DB::table('rooms')->where('team_id', auth()->user()->currentTeam->id)->where('room_types_id', '=', $id)->delete();
+            // DB::table('bookings')->where('team_id', auth()->user()->currentTeam->id)->where('room_id', '=', $room_id)->delete();
+            return true;
+        } catch (\Throwable $th) {
+            return false;
+        }
+    }
+
+    public function removeRoom($id){
+    
+        try {
+            DB::table('rooms')->where('team_id', auth()->user()->currentTeam->id)->where('id', '=', $id)->delete();
             return true;
         } catch (\Throwable $th) {
             return false;
